requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
gevent==23.9.1
greenlet==3.0.3
grequests==0.7.0
idna==3.6
python-decouple==3.8
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.0
zope.event==5.0
zope.interface==6.1
